rename #2 - model -> domain
model은 service, repository, domain을 포괄하는 개념이므로,
의미를 구체화하기 위하여 domain으로 수정합니다.
