+++ src/main/java/domain/Reason.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
