+++ src/main/java/domain/Stadium.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import lombok.*;
