+++ src/main/java/domain/Team.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import lombok.*;
