+++ src/main/java/mapper/StadiumMapper.java
@@ -1,6 +1,6 @@
 package mapper;
 
-import model.Stadium;
+import domain.Stadium;
 
 import java.util.List;
